requirements.txt
pytesseract
Pillow
pdfplumber
tqdm
rapidfuzz
psycopg2-binary
unittest